Drop rc from prerelase tag
The semantic versioning package will compare `rcN` and `rcM`
alphabetically since they start with a character.
This means that `rc5 > rc11 == True`.
If the prerelase component is just digits, they will be compared
numerically and `5 > 11 == False`

The new format is `major.minor.patch-prerelease-ghash`, where
the ghash is a 'g' followed by the 10 first characters of the
commit id.