Also do 'git fetch commit'
in checkout_from_conandata_coordinates.

This is necessary when using [git scm](https://docs.conan.io/2/examples/tools/scm/git/capture_scm/git_capture_scm.html)
integration with Azure DevOps.
When doing a PR, a new commit is done (target branch + PR-branch), but
this commit is not automatically cloned. Probably because it is not
referenced by any branch or tag.

Fetching the commit before checking out the commit fixes that.

I do not know how this works in other DevOps platform. However, if the
commit is already there, this is basically a no-op. It queries the
remote but, if the commit is already there, it does not download anything. So this should be fine for all DevOps platform.